@@ -6,36 +6,35 @@
 
     <groupId>com.xueyi</groupId>
     <artifactId>xueyi</artifactId>
-    <version>2.5.0</version>
+    <version>3.0.0-Alpha1</version>
 
     <name>xueyi</name>
     <url>https://xueyitt.cn</url>
     <description>雪忆管理系统</description>
 
     <properties>
-        <xueyi.version>2.5.0</xueyi.version>
+        <xueyi.version>3.0.0-Alpha1</xueyi.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
         <java.version>17</java.version>
-        <spring-boot.version>2.7.7</spring-boot.version>
-        <spring-cloud.version>2021.0.5</spring-cloud.version>
-        <spring-cloud-alibaba.version>2021.0.4.0</spring-cloud-alibaba.version>
-        <spring-boot-admin.version>2.7.10</spring-boot-admin.version>
-        <mybatis-plus.version>3.5.3</mybatis-plus.version>
+        <spring-boot.version>3.0.1</spring-boot.version>
+        <spring-cloud.version>2022.0.0</spring-cloud.version>
+        <spring-cloud-alibaba.version>2022.0.0.0-RC1</spring-cloud-alibaba.version>
+        <spring-boot-admin.version>3.0.0-M9</spring-boot-admin.version>
+        <mybatis-plus.version>3.5.3.1</mybatis-plus.version>
         <pagehelper.version>5.3.2</pagehelper.version>
-        <netty.version>4.1.67.Final</netty.version>
-        <spring-doc.version>1.6.13</spring-doc.version>
+        <spring-doc.version>2.0.2</spring-doc.version>
         <tobato.version>1.27.2</tobato.version>
         <kaptcha.version>2.3.3</kaptcha.version>
         <druid.version>1.2.15</druid.version>
-        <dynamic-ds.version>3.5.2</dynamic-ds.version>
+        <dynamic-ds.version>3.6.1</dynamic-ds.version>
         <commons.io.version>2.11.0</commons.io.version>
         <commons.fileupload.version>1.4</commons.fileupload.version>
         <velocity.version>2.3</velocity.version>
-        <fastjson2.version>2.0.22</fastjson2.version>
-        <minio.version>8.2.2</minio.version>
-        <poi.version>4.1.2</poi.version>
+        <fastjson2.version>2.0.23</fastjson2.version>
+        <minio.version>8.5.1</minio.version>
+        <poi.version>5.2.3</poi.version>
         <hutool.version>5.8.11</hutool.version>
         <jjwt.version>0.9.1</jjwt.version>
         <lombok.version>1.18.24</lombok.version>
